refactor(config): extract activation and cache instrumentation into dedicated modules (#162)
Behavior-preserving decomposition of config.py (474) / config.ts (377) per STYLE.md single-responsibility. P1.A of the 2.7 refactor program.

- Extract the hidden GameData-generation activation singleton (state+update+query trinity) into activation.py / activation.ts; extract activation-aware cache instrumentation into cache_stats.py (TS already had cacheStats.ts/cacheMetrics.ts). config.py -178 / config.ts -77, leaving only path/env + Config.
- Cycle break: Python one-way DAG (Config.load late-imports activation); TS ESM mutual top-level import (forced -- loadConfig is sync) with a CYCLE INVARIANT comment documenting why a state-extraction split cannot remove it.
- peek_pinned_config/peekPinnedConfig keeps Config.load snapshot-aware; path helpers promoted to public (mirrors TS exports); _current_activation_signature -> current_activation_signature.
- Test rewiring: test_sync_release patch targets -> activation module; TS tests use canonical activation + __resetActivationForTesting (cacheBust no longer isolates activation); bundled-fallback test copies activation.ts.

Both impls green (Python 369 / TS 283). 双轨 CR: independent clean + KHPilot approve; SHOULD-FIX (path-helper promotion, TS cycle) + import-surface CHANGELOG note addressed. Activation/cache symbols moved out of config (see CHANGELOG); MCP tool surface unchanged.
