+++ python/src/prts_mcp/data/images.py
@@ -11,12 +11,9 @@ import logging
 from dataclasses import dataclass
 from typing import Any, Mapping
 
-from prts_mcp.config import (
-    Config,
-    activation_aware_cache,
-    cache_stat,
-    register_activation_listener,
-)
+from prts_mcp.activation import register_activation_listener
+from prts_mcp.cache_stats import activation_aware_cache, cache_stat
+from prts_mcp.config import Config
 from prts_mcp.data.stores import DirectoryStore
 
 _logger = logging.getLogger(__name__)
