+++ ts/src/data/item.ts
@@ -4,7 +4,8 @@
  * Mirrors python/src/prts_mcp/data/item.py.
  */
 
-import { checkActivationChange, loadConfig, registerActivationListener } from "../config.js";
+import { checkActivationChange, registerActivationListener } from "../activation.js";
+import { loadConfig } from "../config.js";
 import { DirectoryStore } from "./stores.js";
 import { CacheMetrics } from "./cacheMetrics.js";
 import type { CacheStat } from "../cacheStats.js";
